jwtauth added to appropriate routes.

+++ routes/api.js
@@ -6,9 +6,12 @@ const api = express.Router();
 const db = require('../models/')
 // const db = require('../models/index');
 
+//middleware to authenticate jwt token
+//jwtauth will attach user propertiy to req upon authentication to all handlebars customization
+var jwtauth = require('./jwtAuth.js');
 
 // get all events
-api.get("/api/events", (req, res) => {
+api.get("/api/events", jwtauth, (req, res) => {
     console.log("/api/events in api.js");
     db.Event.findAll()
         .then(function (data) {
@@ -22,7 +25,7 @@ api.get("/api/events", (req, res) => {
 });
 
 // get one event
-api.get("/api/events/:id", (req, res) => {
+api.get("/api/events/:id", jwtauth, (req, res) => {
 
     db.Event.findOne({
             where: {
@@ -40,7 +43,7 @@ api.get("/api/events/:id", (req, res) => {
 });
 
 // create an event
-api.post("/api/events", (req, res) => {
+api.post("/api/events", jwtauth, (req, res) => {
     //adding in sequelize code here LH
     console.log("/api/events");
     db.Event.create(req.body)
@@ -54,7 +57,7 @@ api.post("/api/events", (req, res) => {
 });
 
 // udpate an event
-api.put("/api/events", (req, res) => {
+api.put("/api/events", jwtauth, (req, res) => {
 
     // console.log("eventID: " + eventID);
 
@@ -69,7 +72,7 @@ api.put("/api/events", (req, res) => {
 });
 
 // delete an event
-api.delete("/api/events/:id", (req, res) => {
+api.delete("/api/events/:id", jwtauth, (req, res) => {
     db.Event.destroy({
         where: {
             id: req.params.id
@@ -80,7 +83,7 @@ api.delete("/api/events/:id", (req, res) => {
 });
 
 // get all users
-api.get("/api/users", (req, res) => {
+api.get("/api/users", jwtauth, (req, res) => {
     console.log("/api/users in api.js");
     db.User.findAll()
         .then(function (data) {
@@ -94,7 +97,7 @@ api.get("/api/users", (req, res) => {
 });
 
 // create a user
-api.post("/api/users", (req, res) => {
+api.post("/api/users", jwtauth, (req, res) => {
 
     db.User.create(req.body)
         .then(result => {
@@ -109,16 +112,18 @@ api.post("/api/users", (req, res) => {
 });
 
 // edit a user
-api.put("/api/users/:id", (req, res) => {
+api.put("/api/users/:id", jwtauth, (req, res) => {
 
 });
 
 // delete a user
-api.delete("/api/users/:id", (req, res) => {
+api.delete("/api/users/:id", jwtauth, (req, res) => {
 
 });
 
 // try to login 
+//NO auth here because this route needs to query if user exists
+//and will then sign a token
 api.post("/api/login", (req, res) => {
     //console.log(req.body.email);
 
@@ -167,7 +172,7 @@ api.post("/api/login", (req, res) => {
 });
 
 // add a volunteer to an event
-api.post("/api/staging", (req, res) => {
+api.post("/api/staging", jwtauth, (req, res) => {
 
 });
